Move exhaustive tests to test-special folder
